Adjusted the location of created exposure-maps
In previous adjustments the now home-created exposure maps
were being written inside the DATA_ARCHIVE dir, which may
mean the local Swift data archive (in case the user has a copy of it).
Well, this was not right because (1) we don't want to overwrite
those files (expomaps) in DATA_ARCHIVE, and (2) we may not have
write permission to do it -- which would cause an error.
Now we are writing the new exposure-maps in our tmpdir (./tmp/expomaps).