Fix the bug in parsing sosta.log
The bug was lying in an old (and bad) parsing of
the line for '+ vignetting ->' flux values.
It assumed that we would have upper-values and negative-fluxes
happening together, while the parsing was correct, the hypothesis
was wrong.
Fixed by doing a better parsing of that line and *not* assuming the
coupling.